adding featured image to subheader. moving breadcrumbs from page to subheader
