+++ wp-content/themes/core/components/routes/page/Page_Controller.php
@@ -6,8 +6,7 @@ namespace Tribe\Project\Templates\Components\routes\page;
 use Tribe\Project\Templates\Components\Abstract_Controller;
 use Tribe\Project\Templates\Components\image\Image_Controller;
 use Tribe\Project\Templates\Components\sidebar\Sidebar_Controller;
-use Tribe\Project\Templates\Components\breadcrumbs\Breadcrumbs_Controller;
-use Tribe\Project\Templates\Models\Breadcrumb;
+
 
 class Page_Controller extends Abstract_Controller {
 
@@ -59,26 +58,6 @@ class Page_Controller extends Abstract_Controller {
 		get_template_part( 'components/document/footer/footer', 'page' );
 	}
 
-	public function render_breadcrumbs(): void {
-		//TODO: let's make this get_breadcrumb_args() and render in template
-		get_template_part(
-			'components/breadcrumbs/breadcrumbs',
-			'page',
-			[ Breadcrumbs_Controller::BREADCRUMBS => $this->get_breadcrumbs() ]
-		);
-	}
-
-	/**
-	 * @return Breadcrumb[]
-	 */
-	protected function get_breadcrumbs(): array {
-		$page = get_the_ID();
-		$url  = $page ? get_permalink( $page ) : home_url();
-
-		return [
-			new Breadcrumb( $url, get_the_title( $page ) ),
-		];
-	}
 
 	public function get_image_args() {
 		if ( ! has_post_thumbnail() ) {
