+++ wp-content/themes/core/components/routes/page/page.php
@@ -10,9 +10,7 @@ $c->render_header();
 
 	<main id="main-content">
 
-		<?php $c->render_breadcrumbs(); ?>
-
-		<?php get_template_part( 'components/header/subheader/subheader' ); ?>
+		<?php get_template_part( 'components/header/subheader/subheader' ); ?> 
 
 		<div class="s-sink t-sink l-sink">
 			<?php the_content(); ?>
